requirements1.txt
xattr>=0.8.0
schedule==0.3.2
SQLAlchemy~=1.1.0b3
Flask-SQLAlchemy~=2.1
Flask-Migrate==2.1.1
redis==2.10.6
requests==2.19.1
pypinyin==0.31.0
pycrypto==2.6.1
scandir==1.7
xlsxwriter>=0.9.3
gunicorn==19.8.1
xlrd==1.0.0
flasgger==0.8.3
emails==0.5.15
celery==4.2.1
ConcurrentLogHandler==0.9.1
eventlet==0.22.1
Flask-AlchemyDumps==0.0.10
Unipath==1.1
amqp==2.3.2
minio
dnspython==1.15.0
consulate==0.6.0
mysqlclient
pymysql==0.8.1
crcmod==1.7
pycryptodome==3.6.6
aliyun-python-sdk-core-v3==2.9.0
aliyun-python-sdk-kms==2.5.0
oss2==2.5.0
qiniu
